refactor(product-test-fixture): apply Dummy and Request differently for the test environment
