+++ src/test/kotlin/ecommerce/dto/DummyProductRequest.kt
@@ -2,7 +2,7 @@ package ecommerce.dto
 
 import java.math.BigDecimal
 
-data class DummyRequest(
+class DummyProductRequest(
     val name: String? = null,
     val price: BigDecimal? = null,
     val imageUrl: String? = null,
